Update Lada.hpp
Added "damageResistance = 0.02099;" to the DZE2 variants.  This should let the armor perform more like a humvee then before (humvee = 0.03099).  This pushes the value of the damage resistance to about 1.5X better.

Reduced fuelcapacity on all upgrade models, as the tank is set @ 50 and not 100.   Because of the lower amount, the tank upgrades shouldn't exceed 105L (to keep it at the correct ratio). However, 150 seems like a fair compromise, 1/2 way between the calculated 105L and the previous 210L.