darkUI fix
In player_updateGUI.sqf, _foodVal and _thirstVal are calculated differently (values closer to 1 are bad) when Dayz_Dark_UI == true. When Dayz_Dark_UI == false (old system), values closer to 0 are bad. This means that when Dayz_Dark_UI == true, health regenerates when both hunger and thirst are LOW instead of high. This fix compensates by redefining the _foodVal and _thirstVal variables for the old UI system, but a better solution might be to rewrite player_updateGUI.sqf to be more compatible.