Remove merge conflict markers
@icomrade - I'm going to start reviewing and testing the commits you've
done here. I'll commit any fixes and additional 1.8.7 stuff that is
needed to this 1.8.7-Rebase branch. Once we have it in a functional
state we can merge it into master if that's good with you.

Let me know if you are working on any files in particular so we don't
duplicate efforts. I'll be in discord while I'm working here too.