Use unmodified 1.8.7 dayz_code stringtable
The only difference between this and the previous version is:
1. Container tags removed
2. Comments added back in
3. str_usract_watch_toggle, str_usract_watch and str_itemwatch are not
commented out. These should not be duplicates if we are using the new
languagecore.xml (according to the comment). If it is a duplicate we
should fix it on 1.8.7 too (after confirming it there).

Using the unmodified file directly from 1.8.7 dayz_code makes it easier
to compare for future updates.