Update more ammo names
See:
https://github.com/DayZMod/DayZ/commit/634214570c2a6c3efe53e2b09bf01250fcc3d2e2

@icomrade - IMO we should use the new blood type system and take out the
config variable. It is not something that can be easily switched between
because it requires manually changing the loot tables and traders which
adds some big files to the mission. Some admins will get confused when
they change the config variable and it doesn't change the loot spawns or
traders.

Admins can make bloodBagONEG abundant on their server if they
don't like it. I think most people will want to use it since it is a
nice extra realism feature and consistent with 1.8.7. We can still leave
the ItemBloodbag class defined in cfgMagazines so admins can go back to
using it if they wish. 1.8.7 even took out the class completely.