Cleanup CfgLoot
There is no point in having CfgBuildingPos.hpp in dayz_epoch_b too. It
was identical to the one in dayz_code and can be included in the mission
from there.

@icomrade couple questions:

1. Why keep a CfgLootLegacy? I know they are very different formats, but
it will just make things messier and more confusing. Also it will be a
PITA to maintain and update two different sets of loot tables in the
long run. I don't think catering to lazy admins is a good enough reason.
It's not that hard to convert to the new format.

2. Why not use the new _DZ weapon classes in the loot tables like
PDW_DZ, AKM_DZ, etc.?

3. Why not use the new box names like DZ_AmmoBoxRU, DZ_AmmoBoxUS,
DZ_MedBox,  etc. in the loot tables?

4. Why remove the attachments from the loot tables?