Fix player_monitor errors
DayzGearSave is no longer used. I will submit a pull to remove it from
vanilla too.

Player_dumpBackpack is not a good solution to stifle duping. Players can
not see friends backpacks dropped on the ground. That is a problem when
they want to trade backpacks. They don't understand why their friend
can't see it. Also players can still get the "gear" scroll action if
they look where the hidden backpack is.

If it is still a problem after all the updates to current dayz code then
we should submit a pull request with a better solution to vanilla.