Correct duplicate object UID check
I forgot numbers greater than 1 million are truncated when converted to
string:

http://killzonekid.com/arma-scripting-tutorials-float-to-string-position-to-string/

KeyStartNumber exceeded 6 digits which is why it was being truncated to
"1e+011" and getting stuck on loop (as that string was already in
currentObjectUIDs) as mentioned in 2df7279.

Starting at 10, there should never be more than a couple duplicates
generated, so keyStartNumber will never exceed about 100 max.  Added an
exit just for a fail safe.