Fix Unsupported language English in stringtable RPT error
Vanilla development commit:

https://github.com/DayZMod/DayZ/commit/b038a1762bc90828c3598af3d06b3e51677b01a1

Fixes RPT error "Unsupported language English (or whatever language you
selected) in stringtable" on load.

We do not have <Original> tags on any other strings in the table, so
there is no reason to use them here. They are not needed. If the game is
set to use a language that does not have an entry it uses the first
available entry for that string (which is English for every string in
the table). Keys should not contain only <Original> tags with no other
languages, as it will cause this error.