Update change log
It is unnecessary to specify default values in configVariables.sqf
comments. All values in this file will be defaults at release. It can't
be edited by admins directly. They need to copy the variables to their
mission to change them, so they will always be able to reference it for
the defaults.