Fix _distance in dze_buildChecks
Passing false as the second parameter to fnc_find_plots always returns
DZE_PlotPole select 0. In dze_buildchecks we want to use DZE_PlotPole
select 1 if the classname is a plot pole.

_distance is already defined as DZE_PlotPole select 0 in the other three
files, so it is redundant to redefine it as the return value.

In player_upgrade.sqf the check for DZE_permanentPlot is not necessary,
because FNC_check_access handles both cases.