Add back Epoch custom lootPos and lootPosSmall
Epoch has custom loot positions for many buildings. Most lootPos have
not changed in vanilla since pre 1.7.x.

The vanilla lootPos are now commented out and  custom Epoch ones are
added back. LootPosSmall positions are now included in lootPos and
marked with //DZE EXTRA

LootPosSmall are included and used by default in 1051, so it is simpler
to add them to the main lootPos lists instead of having a separate loop
in buildingSpawnLoot.sqf for them.