Changed Cinder Wall Strings
The full cinder wall and the half cinder half kit had the same name ingame. I tried to correct it but I saw that the string was too long when you click right on a full cinder wall kit to build it. I renamed both kits so that both kits stand by each other.