Plot for life fixes and typo fixes (#1707)
This fixes the issues I was having with APFL, Basically when downgrading
a locked door the ownerPUID was not getting transferred across so it was
getting a 0 PUID set on it once downgraded so you would then lose access
if there was no plot pole.

Now it correctly gets set and instead of doing 2 PVDZE_Obj_Swaps I am
using an if and else statement.

player_upgrade was also not sending the ownerID into the worldspace
field, this seems to differ from the original APFL script. I also added
in code so that if APFL was enabled and uid is 0 we claim the item with
players UID.

This fixes a typo @ebaydayz introduced to spam the .RPT on startup also
for the shotgun and a typo in the credits.