More Permissive Combat Checks
This commit puts the conditions in which a player is put into combat
more in line with 1.0.5.1. Putting the player in combat for having
another player nearby will ruin building for players that are part of a
group. Putting the player into combat if zombies are nearby would make
building way too frustrating (especially if building on an elevated or
enclosed area where zombies are not a threat) and will attack the player
if not dealt with anyway.