Fix item count being overwritten in AdvTrade right list title
When switching back to category view the item count was overwritten
from:

https://github.com/EpochModTeam/DayZ-Epoch/commit/12a9e9f83edc78c4ec147d86e6b1d77f91323250

It is no longer needed to set it in changeBuySell since calcPrice runs
every time changeBuySell does. The one exception is when switching back
to category view, but in that case it stays in buy mode anyway.

You forgot to add it for single currency as well @oiad