Advanced trading change to buyables list
This changes Advanced trading to change the item color to green if it
matches with an item is in your inventory, in your weapon slots (main
slot and dayz_onBack) is the same backpack and is a vehicle you are
local to.

This brings it in line with default trader, except we're using local
here since that's all that matters.

Thanks to @ebaydayz for the suggestion.