Re-add observer to UH1Y and AH6X
This fixes config errors for unknown animation sources for both
vehicles. The turrets do not have thermal capabilities and this restores
the missing co-pilot/observer seat in the UH1Y. Also I can't seem to get
the turret on the AH6X to rotate, but if someone knows how please leave
a comment or commit the changes. there's currently no overlay for the
observer of the AH6X since it can't move I feel it's limited enough
without an overlay like the UH1Y.