Fix Mozzies/gyrocopters
Fixes wrecks bouncing 500m into the sky, players not dying when crashing
these vehicles, and sound falloff (someone verify in a multiplayer game,
I teleported away from a flying mozzie and it appeared fixed).

I made a new startup and stop sound for the mozzie which is acceptable I
think, but I'm using the MV22 Osprey noise startup and stop noises for
the gyrocopter. The osprey startup is fairly well fitting, but the stop
noice is out of place, if anyone can come up with a better noise please
change it otherwise if it's out of place enough I'll just remove it
altogether.