Color new killfeed death messages
The player who died is colored red and the killer is colored green. The
weapon now points in the direction of the player who died.

Also reverted 66fd5be since it did not help. It is already done globally
in player_death.sqf anyway. I tested again and this actually fixes the
problem.