Fix humanity hit for units in vehicles
The only change is a removal of this check and the corresponding
tabs/spacing:

if (_object == player) then {

That returns false if the player is in a vehicle. It is not needed since
PVCDZ_plr_Humanity EH only fires on the intended player now (sent with
PVC instead of PV).