Fix damage handling of vehicle explosions
Vanilla commit:

https://github.com/DayZMod/DayZ/commit/589791e9f9116e8b230e01d51fabde028298ea7a

This was being falsely triggered before because the distance (3m) was
far too small and it was only counting fast moving vehicles. The player
can be damaged up to 25m away from the vehicle explosion.

The comment was incorrect. It can be easily checked by using the
included diag_log at the top of this file.