Remove animation for switching clothes
If the gear sync script is called during the animation, the player will
fail the gear sync and not change clothes. After removing the anim this
no longer happens. Also I think this is a BS change that only
inconveniences the player and makes something that is already cumbersome
even more so