Allow physics damage from vehicle crash
Vanilla commit:

https://github.com/DayZMod/DayZ/commit/804445830ffa38bba742cc20c6645b75333725f5

This was exiting with 0 damage if _hit=="". It accounts for vehicles not
being damaged at all when crashing full speed into walls. It was also
causing the player to sometimes glitch out of the vehicle or into the
ground when crashing as you noticed before @icomrade.

Also fixed driver getting a humanity hit for damaging passengers in
crash. Reverting 22b2a3a since this is more reliable.