Revert vanilla work bench to player_buildVanilla
Partial revert of:

https://github.com/EpochModTeam/DayZ-Epoch/commit/2d9097a6af14ff21d0cd362eccaf62678c86060e

This workbench is currently not obtainable either by crafting or the
loot tables, so it is not used anyway. But with Epoch player_build
attempting to build it sends the player flying / sliding backwards at
high speed. Using the vanilla player_build that does not happen.

I tested building all the other items changed in the above commit and
none of them have this problem.