Remove unneeded global var Z_AT_FolderLocation
This variable is not useful. If you change it you need to copy all the
AT files to your mission.

It is fewer files in the mission to just overwrite the path in compiles
and fn_selfActions.

Also fixed mistake with not replacing macros in scripts.txt with their
actual values.