Extend DZE_BackpackAntiTheft to support groups
This change allows group members as well as players you are friendly to
to access each others gear.

This also fixes the problem with being able to access peoples gear in
safe zones by just tagging them as a friend, this was quite dangerous
because it required no authorization from the remote player. Now both
players need to be friendly to each other OR in a group (group members
can be considered trust worthy I guess?)

Fixes an issue in fn_check_access where it was trying to get ownerPUID
on a player instead of getPlayerUID