Prevent purchasing negative number of items
Fixes mistake from ffef031 @icomrade . _amount here is the quantity the
player enters in the quantity box when buying. The only time it is less
than or equal to zero is if the player enters zero or a negative number.
It has nothing to do with the buy or sell price.

Thanks Sercan for reporting.