Stop server_publishVehicle3 duping a vehicle if hive fails to update
server_publishVehicle3 first creates the vehicle in the database then
reads it back to see if it's successful, previously it wasn't deleting
the vehicle if it failed.

Added a 1 second delay after the vehicle delete to stop duping of
vehicles also.

Fixed references to server_deleteObj that were using 2 parameters
instead of the required 3 and updated the example.