Update Core Patch BIS_Effects again
Updated to same formatting and comments as latest Corepatch files:

https://github.com/Goliath86/CorePatch/tree/master/CorePatch_FIS/data/scripts

airdestruction.sqf:
- fixed _velocity variable was undefined
- updated, Sa-Matra removed clearVehicleInit
- !isNull checks are not needed because (local objNull) returns false
and (speed objNull) returns zero

BIS_Effects_startEvent:
- 6th parameter is not used in BIS_Effects_Burn, neither in the old nor
new version.

Although the formatting is ugly, it is better to stay consistent with
Bohemia's so they are easy to compare with future updates and
scripts.txt exceptions are the same whether the official or custom files
are used.

Added Sa-Matra's final commit from yesterday:

https://github.com/Goliath86/CorePatch/commit/5e696bebdca1777ef45c64323c6ad38a73667ff2

This partially reverts c15caf5.