Fix gut zombie not showing up reliably
This fixes gut zombie not working correctly, using damage on a zombie to
check if alive only works if the zombie takes a large large hit. it
works great for checking  vehicles or objects but not zombies/players,
need to use alive in this situation.

Thanks to @theduke77 for the heads up.