Block another A2OA script execution exploit
Thanks to Dihan for reporting on Discord.

This exploit may work with other event handlers, displays and controls.
Please leave a comment if you find any others that work. I tested GPS,
abort, MP setup and group menu displays, as well as the map diary list
box controls, but none of those had this problem.

This should be fixed with an A2OA patch, please report to Bohemia.