Removed "=" from dze_buildChecks.sqf
Salival recommended a >=, however upon further thought, this would be irrelevant as this variable is meant to prevent building over a certain height, not at that specific height. Having >= would restrict building at exactly 30 meters for example rather than simply above it. The = sign was also not used in modular_build.sqf