Revert "Only allow lock removal if door is unlocked"
This reverts commit b8dba25

As discussed with oiad on Discord, this serves no purpose now, since
DZE_lockDoor == charID means the player either just unlocked, built,
upgraded or downgraded the door, so they want to see the action.
Logically a combo lock can be removed while a door is open. In case of
it being locked, you already know the code, so it is assumed you put it
in before removing it. Completing two-three steps at once is more
convenient for the player.

Also removed _isDestructable because there are no classes which inherit
from "BuiltItems" and have a lock