Make Hide Body customizable
Do not use hideBody as command without the action. It just hides the body but it still can be geared. If using hideBody in correlation with deleteVehicle it breaks the allDead command.

player action ["hideBody", body] deletes the body and the marker from the map but it stays inside allDead.

This fixes the issue that hided players were still shown on the map and still searched for their body.

Needs additional testing for a new release.