Adding AK-107 to attachment system, loot, trader
This adds 6 AK-107 versions to the dayz attachment system. There is no SD version at the moment.

The RPK-74 magazine could only be used with the AK-74 SD. This was unintentional. All AK-74 versions support the RPK-74 magazine now.

The RPK-74 and the AK-74 magazine description needed a few adjustments.