Fix disappearing weapons from backpack
The normal intention to have a bag with this small size is, that not a weapon can be carried. Having 5 slots or more in a backpack is normally the signal that at least a secondary weapon can be carried inside this backpack. If a weapon will be put inside a bag without any free weapon slot, the weapon just disappears. To have a better and more intuitive playstyle every backpack with more as 5 slots should be able to carry at least 1 weapon.