Move no longer used maps to archive.
All files in the archive will no be updated. Ive added a copy of the current trader files too, so the old traders and trader cities can still be used with the current trader files.

Give me a hint if you think some of the files should be moved back.