Clean up CH53 configs
In consultation with Seelenapparat we changed the configs from the ch53. This removes strop down and up, the collect and detach cargo options. both actions are not really used in epoch and needed an init script which runs permanently while in the helicopter. It is the same for the sfx sounds from this helicopter. In this special case we decided that performance is more valuable as the removed actions. If someone uses the removed actions in scripts or something like that please message me on discord.