Adding german medical translations, minor changes
Changing the messages for wipes. It cannot be done like it was before for the german language.
Thx @RedLink-kam for the russain translations.

Ive re-ordered the new strings so the Namalsk compatibility is the last again. Namalsk should be always last since it is just an addon from a different mod.