Update tents
equip_tent_poles are now in the loot tables and can be used for tent crafting.
The recipe for crafting tents and nettings has changed a bit. Instead of poles, tent poles are needed now.
The string for equip_tent_poles is confusing if the requirements are 6 Tent Poles but meant is 1x (6 Tent Poles).